introduce portable test.Chmod()
