+++ hcl/fmt_test.go
@@ -1432,6 +1432,38 @@ func TestFormatTreeReturnsEmptyResultsForEmptyDir(t *testing.T) {
 	assert.EqualInts(t, 0, len(got), "want no results, got: %v", got)
 }
 
+func TestFormatTreeFailsOnNonAccessibleSubdir(t *testing.T) {
+	const subdir = "subdir"
+	tmpdir := t.TempDir()
+	test.Mkdir(t, tmpdir, subdir)
+
+	assert.NoError(t, os.Chmod(filepath.Join(tmpdir, subdir), 0))
+	defer func() {
+		assert.NoError(t, os.Chmod(filepath.Join(tmpdir, subdir), 755))
+	}()
+
+	_, err := hcl.FormatTree(tmpdir)
+	assert.Error(t, err)
+}
+
+func TestFormatTreeFailsOnNonAccessibleFile(t *testing.T) {
+	const filename = "filename.tm"
+
+	tmpdir := t.TempDir()
+	test.WriteFile(t, tmpdir, filename, `globals{
+	a = 2
+		b = 3
+	}`)
+
+	test.Chmod(t, filepath.Join(tmpdir, filename), 0)
+	defer func() {
+		test.Chmod(t, filepath.Join(tmpdir, filename), 0755)
+	}()
+
+	_, err := hcl.FormatTree(tmpdir)
+	assert.Error(t, err)
+}
+
 func TestFormatTreeFailsOnNonExistentDir(t *testing.T) {
 	tmpdir := t.TempDir()
 	_, err := hcl.FormatTree(filepath.Join(tmpdir, "non-existent"))
