+++ test/fs.go
@@ -15,6 +15,7 @@
 package test
 
 import (
+	"io/fs"
 	"os"
 	"path/filepath"
 	"testing"
@@ -68,4 +69,8 @@ func AssertFileContentEquals(t *testing.T, fname string, want string) {
 	if diff := cmp.Diff(string(got), string(want)); diff != "" {
 		t.Fatalf("-(%s) +(%s):\n%s", got, want, diff)
 	}
+}
+
+func Chmod(t *testing.T, fname string, mode fs.FileMode) {
+	assert.NoError(t, chmod(fname, mode))
 }
